src/third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/module/dynamic-import/alpha/base-url-worker-importScripts-expected.txt

This is a testharness.js-based test.
FAIL Relative URL-like from same-origin importScripts() assert_equals: expected "gamma/import.js" but got "beta/import.js"
PASS Absolute URL-like from same-origin importScripts()
FAIL Relative URL-like from cross-origin importScripts() assert_equals: Relative URL-like specifier resolution should fail expected "(unreached)" but got "beta/import.js"
PASS Absolute URL-like from cross-origin importScripts()
Harness: the test ran to completion.
